containers配下のrootとindexをsrc直下のsetupとappにファイル名変更＆移動
