+++ index.ios.js
@@ -1,17 +1,8 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
- */
+'use strict';
 
-import React, { Component } from 'react';
 import {
   AppRegistry,
-  StyleSheet,
-  Text,
-  View
 } from 'react-native';
-import Root from './src/containers/root';
+import Root from './src/setup';
 
 AppRegistry.registerComponent('RecoSpo', () => Root);
-
